+++ app/(front_office)/(routes)/layout.tsx
@@ -6,7 +6,7 @@ interface Props {
 
 const Layout = ({ children } : Props) => {
   return (
-    <div className='max-w-7xl mx-auto px-4 sm:px-6 lg:px-8'>
+    <div className='min-h-screen max-w-7xl mx-auto px-4 sm:px-6 lg:px-8'>
       {children}
     </div>
   )
