+++ app/(front_office)/(routes)/offres-lists/[id]/postuler/page.tsx
