+++ app/(front_office)/page.tsx
@@ -1,12 +1,9 @@
-import Hero from '@/components/front_office/Hero'
 import React from 'react'
 
 const Home = () => {
   return (
-    <div className='space-y-6 max-w-7xl mx-auto px-4 sm:px-6 lg:px-8'>
-      <Hero />
-    </div>
+    <div className='min-h-screen max-w-7xl mx-auto px-4 sm:px-6 lg:px-8'>Home</div>
   )
 }
 
-export default Home;
+export default Home
